Address some review comments + make new clippy happy
