+++ artifacts/checksums.txt
@@ -13,11 +13,11 @@ c3c8b876ab02edc59110580470a0840974bc37ca1cfe04f895a23fa30b98614a  cctp_adapter.w
 dfe70e515ce38d6cb44f3a8b9d398f82988524f58e1ba8eb21f4bff297c7e7bb  marketplace.wasm
 24fe99f9b50a9e35466d55377e6db9056304dcb5511c7cec871bfd607165adbc  mars_adapter.wasm
 986882765a6e6eba70796adc077b7d1848a67ce410e270624fe23e5536f2ae1e  proxy.wasm
-f777d53a5d55741cfc4993bc3af622debaee8e724cbc3fe746ebb7c6d9b87929  shares_converter.wasm
+a99cc61b7cde60afc79619e241ea7b064df3cb357305b416c5af5900902c0754  shares_converter.wasm
 1b885c15bbe9de8ca8d2f13e32f8736787a1ac585d83c3e1000cb8c4ac6f1030  skip_adapter.wasm
 c6be8e83fde7f61b529575a99a650e4684b3cd15c1fc144a8af787f03c985560  skip_adapter_cosmoshub.wasm
 7ccb9370aeadb7b651592713f7dc13747896e72f8a19a933a40ae541b7137c14  st_token_info_provider.wasm
 ea2c1da6dbedacd8f32b098905dd5790b095075a1a7457bfbf123c86465987eb  tribute.wasm
 f904e2d69a8c43984590ddba3855079754d09a7e09b76c1557d89d6e96e8199d  user_registry.wasm
 1802c554f72d444251a38621d6984ff225b935940c72ab015b5235024eff2df7  vault.wasm
-0c2d1e114b2f9b5ca026f0df12c9e81be1595637aafbf11d13721f10d2b7e333  vault_cosmoshub.wasm
+54044213bd5f7f88ae670019af096e68441a3b3228b6609085629f678d3cbe61  vault_cosmoshub.wasm
