+++ contracts/inflow/shares-converter/src/lib.rs
@@ -2,3 +2,6 @@ pub mod contract;
 pub mod error;
 pub mod msg;
 pub mod state;
+
+#[cfg(test)]
+mod testing;
